PythonProject/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
cryptography==41.0.8
python-multipart==0.0.6
